Rename Tagged → Unique, move to own file
